+++ drupal-mybatis/trunk/src/main/java/org/gbif/user/mybatis/UserRoleTypeHandler.java
@@ -3,7 +3,7 @@ package org.gbif.user.mybatis;
 import org.gbif.api.model.vocabulary.UserRole;
 import org.gbif.mybatis.type.BaseEnumTypeHandler;
 
-public class UserRoleTypeHandler extends BaseEnumTypeHandler<Integer, UserRole> {
+public class UserRoleTypeHandler extends BaseEnumTypeHandler<Long, UserRole> {
 
   public UserRoleTypeHandler() {
     super(new UserRoleTypeConverter());
